Add http-01 support to the Let's Encrypt mechanism.
Since the autocert.Manager is always created regardless of whether
Let's Encrypt is being used or not I opted to always wrap the HTTP
redirect handler with the autocert.Manager HTTPHandler. This means
that port 80 will always be available to handle a Let's Encrypt challenge
as well as do the redirect. This should get Let's Encrypt support
functioning again since tls-sni-02 and tls-sni-01 were deprecated.